CAMEL-21858: Moving httpMethod parameter handling to HttpSpanDecorator
