+++ components/camel-tracing/src/test/java/org/apache/camel/tracing/decorators/AbstractHttpSpanDecoratorTest.java
@@ -20,7 +20,6 @@ import org.apache.camel.Endpoint;
 import org.apache.camel.Exchange;
 import org.apache.camel.Message;
 import org.apache.camel.tracing.MockSpanAdapter;
-import org.apache.camel.tracing.SpanDecorator;
 import org.apache.camel.tracing.TagConstants;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -33,11 +32,21 @@ public class AbstractHttpSpanDecoratorTest {
 
     private static final String TEST_URI = "http://localhost:8080/test";
 
-    private HttpSpanDecorator decorator;
+    private AbstractHttpSpanDecorator decorator;
 
     @BeforeEach
     public void before() {
-        this.decorator = new HttpSpanDecorator();
+        this.decorator = new AbstractHttpSpanDecorator() {
+            @Override
+            public String getComponent() {
+                return null;
+            }
+
+            @Override
+            public String getComponentClassName() {
+                return null;
+            }
+        };
     }
 
     @Test
@@ -49,18 +58,6 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(exchange.getIn()).thenReturn(message);
         Mockito.when(message.getHeader(Exchange.HTTP_METHOD)).thenReturn("PUT");
         Mockito.when(endpoint.getEndpointUri()).thenReturn("http://localhost:8080/endpoint");
-
-        SpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
 
         assertEquals("PUT", decorator.getOperationName(exchange, endpoint));
     }
@@ -159,18 +156,6 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(exchange.getIn()).thenReturn(message);
         Mockito.when(message.getHeader(Exchange.HTTP_URI, String.class)).thenReturn(TEST_URI);
 
-        SpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
-
         MockSpanAdapter span = new MockSpanAdapter();
 
         decorator.pre(span, exchange, endpoint);
@@ -190,18 +175,6 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(message.getHeader(Exchange.HTTP_URI, String.class)).thenReturn("Another URL");
         Mockito.when(message.getHeader(Exchange.HTTP_URL, String.class)).thenReturn(TEST_URI);
 
-        AbstractHttpSpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
-
         assertEquals(TEST_URI, decorator.getHttpURL(exchange, endpoint));
     }
 
@@ -215,18 +188,6 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(exchange.getIn()).thenReturn(message);
         Mockito.when(message.getHeader(Exchange.HTTP_URI, String.class)).thenReturn(TEST_URI);
 
-        AbstractHttpSpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
-
         assertEquals(TEST_URI, decorator.getHttpURL(exchange, endpoint));
     }
 
@@ -238,18 +199,6 @@ public class AbstractHttpSpanDecoratorTest {
 
         Mockito.when(endpoint.getEndpointUri()).thenReturn(TEST_URI);
         Mockito.when(exchange.getIn()).thenReturn(message);
-
-        AbstractHttpSpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
 
         assertEquals(TEST_URI, decorator.getHttpURL(exchange, endpoint));
     }
@@ -263,18 +212,6 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(endpoint.getEndpointUri()).thenReturn("netty-http:" + TEST_URI);
         Mockito.when(exchange.getIn()).thenReturn(message);
 
-        AbstractHttpSpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
-
         assertEquals(TEST_URI, decorator.getHttpURL(exchange, endpoint));
     }
 
@@ -286,54 +223,11 @@ public class AbstractHttpSpanDecoratorTest {
         Mockito.when(exchange.getMessage()).thenReturn(message);
         Mockito.when(message.getHeader(Exchange.HTTP_RESPONSE_CODE, Integer.class)).thenReturn(200);
 
-        SpanDecorator decorator = new AbstractHttpSpanDecorator() {
-            @Override
-            public String getComponent() {
-                return null;
-            }
-
-            @Override
-            public String getComponentClassName() {
-                return null;
-            }
-        };
-
         MockSpanAdapter span = new MockSpanAdapter();
 
         decorator.post(span, exchange, null);
 
         assertEquals(200, span.tags().get(TagConstants.HTTP_STATUS));
-    }
-
-    @Test
-    public void testMethodInHttpMethodParam() {
-        Endpoint endpoint = Mockito.mock(Endpoint.class);
-        Exchange exchange = Mockito.mock(Exchange.class);
-        Message message = Mockito.mock(Message.class);
-
-        Mockito.when(endpoint.getEndpointUri()).thenReturn("http://localhost:8080/endpoint?httpMethod=POST");
-        Mockito.when(exchange.getIn()).thenReturn(message);
-        Mockito.when(message.getHeader(Exchange.HTTP_URI, String.class))
-                .thenReturn("http://localhost:8080/endpoint?httpMethod=POST");
-
-        assertEquals(AbstractHttpSpanDecorator.POST_METHOD,
-                decorator.getHttpMethod(exchange, endpoint));
-    }
-
-    @Test
-    public void testMethodInHttpMethodParamUsingHeader() {
-        Endpoint endpoint = Mockito.mock(Endpoint.class);
-        Exchange exchange = Mockito.mock(Exchange.class);
-        Message message = Mockito.mock(Message.class);
-
-        Mockito.when(endpoint.getEndpointUri()).thenReturn("http://localhost:8080/endpoint?httpMethod=POST");
-        Mockito.when(exchange.getIn()).thenReturn(message);
-        Mockito.when(message.getHeader(Exchange.HTTP_METHOD)).thenReturn(HttpMethods.GET);
-        Mockito.when(message.getHeader(Exchange.HTTP_URI, String.class))
-                .thenReturn("http://localhost:8080/endpoint?httpMethod=POST");
-
-        assertEquals(AbstractHttpSpanDecorator.POST_METHOD,
-                decorator.getHttpMethod(exchange, endpoint));
     }
 
 }
